requirements.txt
# Core RL & Deep Learning
gymnasium==0.29.1
torch==2.2.0
numpy==1.26.4
scipy==1.12.0

# AWS Integration
boto3==1.34.34
botocore==1.34.34

# Data Processing
pandas==2.2.0
pyarrow==15.0.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
tensorboard==2.16.2
plotly==5.19.0

# Config Management
pyyaml==6.0.1

# API Serving
fastapi==0.111.1
uvicorn==0.30.1
pydantic==2.7.4

# Dashboard
streamlit==1.36.0

# Experiment Tracking
mlflow==2.14.3

# Utilities
tqdm==4.66.2
python-dotenv==1.0.1

# Development Tools
pytest==8.0.0
black==24.2.0

# Stable Baselines3 (optional: for comparison)
stable-baselines3==2.2.1
